# api/openapi.yaml
openapi: 3.0.0
info:
  title: RPG Game API
  version: 1.0.0
  description: API for managing RPG game sessions, battles, and loot

servers:
  - url: /api/v1
    description: API v1

tags:
  - name: games
    description: Game management
  - name: players
    description: Player management
  - name: enemies
    description: Enemy management
  - name: loots
    description: Loot management

paths:
  /games:
    get:
      tags:
        - games
      summary: Get all games
      operationId: getGames
      responses:
        '200':
          description: Games retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/top:
    get:
      tags:
        - games
      summary: Get top games
      operationId: getTopGames
      responses:
        '200':
          description: Top games retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '404':
          description: Games not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/new:
    post:
      tags:
        - games
      summary: Start a new game session
      operationId: createGame
      requestBody:
        description: Number of enemies to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enemies:
                  type: number
                  example: 10
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - games
      summary: Get game status
      operationId: getGame
      responses:
        '200':
          description: Game status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - games
      summary: Delete a game
      operationId: deleteGame
      responses:
        '200':
          description: Game deleted successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}/attack:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      tags:
        - games
      summary: Player attacks an enemy
      operationId: attackEnemy
      responses:
        '200':
          description: Attack result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
        '400':
          description: Invalid attack
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}/defend:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      tags:
        - games
      summary: Player defends against an attack
      operationId: defendAttack
      responses:
        '200':
          description: Defense result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
        '400':
          description: Invalid defense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}/escape:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      tags:
        - games
      summary: Player attempts to escape
      operationId: escapeBattle
      responses:
        '200':
          description: Escape attempt result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
        '400':
          description: Invalid escape attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}/end:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      tags:
        - games
      summary: End the game
      operationId: endGame
      responses:
        '200':
          description: Game ended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}/enemy:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - games
      summary: Get the current enemy in a game
      operationId: getCurrentEnemy
      responses:
        '200':
          description: Current enemy retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enemy'
        '404':
          description: Game or enemy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - games
      summary: Add an enemy to a game
      operationId: addEnemyToGame
      responses:
        '200':
          description: Enemy added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enemy'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}/enemy/{enemyId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: enemyId
        in: path
        required: true
        schema:
          type: integer
    delete:
      tags:
        - games
      summary: Delete the enemy from a game
      operationId: removeEnemyFromGame
      responses:
        '200':
          description: Current enemy deleted successfully
        '404':
          description: Game or enemy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}/loot:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - games
      summary: Get loots earned in the game
      operationId: getGameLoot
      responses:
        '200':
          description: Game loot retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loot'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games/{id}/loot/{lootId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: lootId
        in: path
        required: true
        schema:
          type: integer
    delete:
      tags:
        - games
      summary: Delete loot from game
      operationId: removeLootFromGame
      responses:
        '200':
          description: Current loot deleted successfully
        '404':
          description: Game or loot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /players:
    get:
      tags:
        - players
      summary: Get all players
      operationId: getPlayers
      responses:
        '200':
          description: Players retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '404':
          description: Players not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /players/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - players
      summary: Get player details
      operationId: getPlayer
      responses:
        '200':
          description: Player details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - players
      summary: Delete a player
      operationId: deletePlayer
      responses:
        '200':
          description: Player deleted successfully
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /players/new:
    post:
      tags:
        - players
      summary: Create a new player
      operationId: createPlayer
      requestBody:
        description: The name of the player
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "John Doe"
      responses:
        '201':
          description: Player created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /enemies:
    get:
      tags:
        - enemies
      summary: Get all enemies
      operationId: getEnemies
      responses:
        '200':
          description: Enemies retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Enemy'
        '404':
          description: Enemies not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /enemies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - enemies
      summary: Get enemy details
      operationId: getEnemy
      responses:
        '200':
          description: Enemy details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enemy'
        '404':
          description: Enemy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /loots:
    get:
      tags:
        - loots
      summary: Get all loots
      operationId: getLoots
      responses:
        '200':
          description: Loots retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loot'
        '404':
          description: Loots not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /loots/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - loots
      summary: Get loot details
      operationId: getLoot
      responses:
        '200':
          description: Loot details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loot'
        '404':
          description: Loot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Loot:
      description: A loot item in the game
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the loot
          minimum: 1
        name:
          type: string
          description: The name of the loot
          minLength: 1
        value:
          type: integer
          description: The value of the loot
          minimum: 0

    Character:
      description: A character in the game
      type: object
      properties:
        id:
          description: The unique identifier for the character
          type: integer
          minimum: 1
        name:
          description: The name of the character
          type: string
          minLength: 1
        health:
          description: The current health of the character
          type: integer
          minimum: 0
        attackPower:
          description: The attack power of the character
          type: integer
          minimum: 0
        luck:
          description: The luck of the character
          type: integer
          minimum: 0
          maximum: 100
        level:
          description: The level of the character
          type: integer
          minimum: 1

    Player:
      description: A player in the game
      type: object
      properties:
        character:
          $ref: '#/components/schemas/Character'
        experience:
          description: The experience points of the player
          type: integer
          minimum: 0
        levelUpExperience:
          description: The experience points required to level up
          type: integer
          minimum: 0

    Enemy:
      description: An enemy in the game
      type: object
      properties:
        character:
          $ref: '#/components/schemas/Character'
        loot:
          $ref: '#/components/schemas/Loot'

    Game:
      description: A game in the game
      type: object
      properties:
        id:
          description: The unique identifier for the game
          type: integer
          minimum: 1
        player:
          description: The player in the game
          $ref: '#/components/schemas/Player'
        score:
          description: The score of the game
          type: integer
          minimum: 0
        createdAt:
          description: The date and time the game was created
          type: string
          format: date-time
        enemies:
          description: The enemies in the game
          type: array
          items:
            $ref: '#/components/schemas/Enemy'
        loot:
          description: The loot in the game
          type: array
          items:
            $ref: '#/components/schemas/Loot'

    ActionResult:
      description: The result of an action in the game
      type: object
      properties:
        player:
          description: The player in the game
          $ref: '#/components/schemas/Player'
        enemy:
          description: The enemy in the game
          $ref: '#/components/schemas/Enemy'
        message:
          description: The message of the action
          type: string

    Error:
      description: An error in the game
      type: object
      properties:
        code:
          description: The code of the error
          type: string
        message:
          description: The message of the error
          type: string